st/watchify-based bench scripts
