+++ .gitignore
@@ -10,3 +10,4 @@ coverage
 .DS_Store
 .nyc_output
 /debug/access-token.js
+/bench/bench.js
