+++ bench/index.html
@@ -12,7 +12,7 @@
 
 <body>
     <div id="benchmarks"></div>
-    <script src="/bench/index.js"></script>
+    <script src="bench.js"></script>
 </body>
 
 </html>
