+++ package.json
@@ -44,7 +44,6 @@
     "babelify": "^7.3.0",
     "benchmark": "~2.1.0",
     "browserify": "^13.0.0",
-    "browserify-middleware": "^7.0.0",
     "clipboard": "^1.5.12",
     "concat-stream": "1.5.1",
     "coveralls": "^2.11.8",
@@ -55,7 +54,6 @@
     "eslint": "^2.5.3",
     "eslint-config-mourner": "^2.0.0",
     "eslint-plugin-html": "^1.5.1",
-    "express": "^4.13.4",
     "gl": "^4.0.1",
     "handlebars": "4.0.5",
     "highlight.js": "9.3.0",
@@ -101,7 +99,12 @@
     "build-min": "browserify js/mapbox-gl.js --debug --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
     "watch-min": "watchify js/mapbox-gl.js --debug --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl -o dist/mapbox-gl.js -v",
     "build-token": "browserify debug/access-token-src.js --debug --transform envify > debug/access-token.js",
-    "start": "npm run build-token && npm run watch-dev & st -nc -p 9966 -i index.html . & wait",
+    "build-bench": "browserify bench/index.js --debug --t [babelify --presets react] -t unassertify -t envify > bench/bench.js",
+    "watch-bench": "watchify bench/index.js --debug --t [babelify --presets react] -t unassertify -t envify -o bench/bench.js -v",
+    "download-bench-data": "node bench/download-data.js",
+    "start-server": "st -nc -p 9966 -i index.html .",
+    "start": "npm run build-token && npm run watch-dev & npm run start-server & wait",
+    "start-bench": "npm run download-bench-data && npm run build-token && npm run watch-bench & npm run start-server & wait",
     "build-docs": "documentation build --github --format html -c documentation.yml --theme ./docs/_theme --output docs/api/",
     "build": "npm run build-docs # invoked by publisher when publishing docs on the mb-pages branch",
     "start-docs": "npm run build-min && npm run build-docs && jekyll serve -w",
